perf: move ServiceUsageStatus to a separate file
